[ES-288][fix][driver][MMC] 1.add fatfs mmc func call error log for easy debug 2.fix mmc example bus width from 8 to 4 bit for demo board 3.add emmc high speed switch error log for easy debug mio set wrong

+++ 3rdparty/lib/FATFS/diskio.h
@@ -16,7 +16,7 @@ extern "C" {
 
 
 /* Status of Disk Functions */
-typedef BYTE	DSTATUS;
+typedef DWORD	DSTATUS;
 
 /* Results of Disk Functions */
 typedef enum {
